Split repository.ts: extract member projects, retiring a layout-coupled test (8/N)
Eighth slice of the repository.ts carve-up (audit L14). Moves the member-projects
queries — the self-declared project showcase behind share_project/list_projects
(#646) and its interests cross-reference (#718) — into
repository/memberProjects.ts. All 8 moved symbols verified byte-identical.

This one was held back from earlier batches because tests/tools.test.ts asserted
the `// --- Member projects` banner exists IN repository.ts, slicing the region
between that banner and the next to prove the member-projects queries make no
outbound HTTP call (a stored project link is DATA and must never be fetched).
Moving the section would have made that indexOf fail — the exact break this
campaign already hit once when #804 moved `// --- Member notes` out from under
the region's end.

Rather than re-point the slice at another banner, the extraction removes the
need for slicing at all: the region is now a FILE, so the test scans
repository/memberProjects.ts wholesale. That is simpler and strictly stronger —
it no longer depends on the file's textual layout at all, and a module boundary
cannot drift the way a comment delimiter can. The SECURITY: intent is unchanged
and the module carries a header stating the invariant the test enforces.

repository.ts is now 5,096 lines, down from 7,137 at the start of the campaign.
